renaming statusbar file to be more obviously a component
